Changed parseInt to parseFloat for decimals to work right.

+++ sites/calc.js
@@ -7,104 +7,88 @@ var displayMessage = '';
 
 /***********
 functions to
-update the 
-display 
+update the
+display
 ***********/
 
 function displayZero() {
 	displayMessage += 0;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayOne() {
 	displayMessage += 1;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayTwo() {
 	displayMessage += 2;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayThree() {
 	displayMessage += 3;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayFour() {
 	displayMessage += 4;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayFive() {
 	displayMessage += 5;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displaySix() {
 	displayMessage += 6;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displaySeven() {
 	displayMessage += 7;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayEight() {
 	displayMessage += 8;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayNine() {
 	displayMessage += 9;
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayDecimal() {
 	displayMessage += '.';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayModulus() {
 	displayMessage += '%';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 
 function displayAdd() {
 	displayMessage += '+';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displaySubtract() {
 	displayMessage += '-';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayMultiply() {
 	displayMessage += '*';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
 function displayDivide() {
 	displayMessage += '/';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
@@ -112,28 +96,27 @@ function displayEquals() {
 	var resultArray, result;
 	if (displayMessage.includes('+')) {
 		resultArray = displayMessage.split('+');
-		result = parseInt(resultArray[0]) + parseInt(resultArray[1]);
+		result = parseFloat(resultArray[0]) + parseFloat(resultArray[1]);
 	} else if (displayMessage.includes('-')) {
 		resultArray = displayMessage.split('-');
-		result = parseInt(resultArray[0]) - parseInt(resultArray[1]);
+		result = parseFloat(resultArray[0]) - parseFloat(resultArray[1]);
 	} else if (displayMessage.includes('*')) {
 		resultArray = displayMessage.split('*');
-		result = parseInt(resultArray[0]) * parseInt(resultArray[1]);
+		result = parseFloat(resultArray[0]) * parseFloat(resultArray[1]);
 	} else if (displayMessage.includes('/')) {
 		resultArray = displayMessage.split('/');
-		result = parseInt(resultArray[0]) / parseInt(resultArray[1]);
+		result = parseFloat(resultArray[0]) / parseFloat(resultArray[1]);
 	} else {
 		resultArray = displayMessage.split('%');
-		result = parseInt(resultArray[0]) % parseInt(resultArray[1]);
+		result = parseFloat(resultArray[0]) % parseFloat(resultArray[1]);
 	}
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent += ' = '+result;
 }
 
 
 /*****************
-event listener for 
+event listener for
 button clicks:
 0-9, ., +, -, /, *, =
 ******************/
@@ -198,7 +181,6 @@ event listeners
 
 function clearAll() {
 	displayMessage = '';
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
@@ -208,7 +190,6 @@ clearEl.addEventListener('click', clearAll, false);
 
 function backSpace() {
 	displayMessage = displayMessage.substring(0, (displayMessage.length-1));
-
 	var elMsg = document.getElementById('answer');
 	elMsg.textContent = displayMessage;
 }
